Refactor chaining scope operators for PHP 7.x
