+++ src/main/php/lang/ast/emit/PHP74.class.php
@@ -13,6 +13,7 @@ class PHP74 extends PHP {
     ArrayUnpackUsingMerge,
     AttributesAsComments,
     CallablesAsClosures,
+    ChainScopeOperators,
     MatchAsTernaries,
     NonCapturingCatchVariables,
     NullsafeAsTernaries,
@@ -20,7 +21,6 @@ class PHP74 extends PHP {
     OmitConstantTypes,
     ReadonlyClasses,
     RewriteBlockLambdaExpressions,
-    RewriteClassOnObjects,
     RewriteEnums,
     RewriteExplicitOctals,
     RewriteProperties,
